feat(codegen): make pc a module
